Huge UI, Content and code update
Added so much!

Lots of artwork has been added, but sadly with lots of placeholders. Oh well, they'll be replaced when the design team do some stuff.

---UI---
-Reworked the Main Menu layout
-Designed some buttons
-Main Menu has nice animations now

---FEATURES---
-Added some models
-Streamlined accessing databases by limiting how many are used
-Fixed scene transitions
-Set up lots of objects for playing sounds, but not fully functioning yet

---GALLERY---
-Gallery page now exists
-Added some cool new scrolly business (vertical scrolling content)
-Added Scrollbar, but will disable function, keeping looks only
-Added buffers at the top and bottom to keep things looking nice

--FILES---
Restructured the project massively with the introduction of the Resources folder. This is just to streamline the repo/folders and it seems to be nicer so it's a win.